add support for lists in max
